Added windows function to open a directory Moved some internal shaders/textures into the internal/ directory. Still missing some though Added ability to change asset directory in Erebor Added ability to change title of windows

+++ Mahakam/src/Mahakam/Renderer/Mesh.cpp
@@ -364,7 +364,7 @@ namespace Mahakam
 								// TODO: The spec is vague on when to do 'inv(parentMat) * nodeMat * invBindMatrix'
 								// For now it is done inside the Scene, once per frame
 								skinnedMesh->NodeHierarchy[nodeIndex[nodeID]].offset = invMatrices[i]; // Override offset to be the bone's inverse matrix
-								skinnedMesh->BoneInfoMap[node.name] = i;
+								skinnedMesh->BoneInfoMap[node.name] = (int)i;
 							}
 						}
 					}
